use rewritten device
in the rewrite for Ohua we just needed to add methods and some
types to the Device Trait. Those methods are combinations of exiting
methods i.e. they have a default implementation in the Trait.
So we can "tune" the existing device for
gem5 by implementing this rewritten Device Trait instead of the smoltcp
original. This commit does so, for now without using the additional
methods.