start adding interface and socket
Start moving the required and rewritten parts of the interface and the tcp socket to
m3. The plan is to import as much functionality as possible from smoltcp and only
reimplement the parts that are either actually rewritten, or part of the implemementation
e.g. for the interface and must therefor be present to be able to replace the
interface of smoltpc by our version in m3.