using smoltcp local
We assume an initial setting for Ohua, where the "app" and the code of it's
components are available as input. So we use smoltcp as a subdir of M3, instead of importing it. This also has the advantage, that we can directly use M3 implementation of device/socket and heap-allocated types in the rewrite