impl rewritten device trait
To use m3s device dirvrs with the rewritten smoltcp version we needed to add
implementations for the (slightly) rewritten device trait.
As we want to keep using the m3 networking as is for the client later, we do
not replace the original device trait but add the new one.