minimal client + build xml
To build our server client dialouge for the key-value-store case we need a client app connected to the 'normal' net server, that would then talk to our smoltcp-server. The current version is an adapted version of the client from
the ycsb benchmark case. (By adapted I mean I've cut anything except instantiating a socket, sending a default request and trying to receive a response)