xml file working - communication still not
The pager app is needed to ave a semaphor among client and server, such that the client would only try to connect after the server started to listen. In turn the pager requires the m3fs to work.

And we needed to
a) specify that smotlcp_server and client can use the semaphor <sem name="net">
b) specify that the server can open a device tile (<tiles type="nicdev" />)

Still the communication among smoltcp_server and client is broken. Server gets an empty packet, but seemingly not from the client and response doesn't get through.