store handles operation length
To circumvent smoltcp's limitiation in dynamic stream object handling
m3 uses the mechanism to
a) keep one constant connection among clinet and server and
b) separate application data by inside th stream by sending a length message
   in advance of every request and response

To keep the Rust-C++ interface minimal, we want to refactor the 'store' to a
wrapper for LvLDB calls. So this implements the 'parsing' of the client
request stream into length infos and 'operations'.
Operations shall be passed to the lebeldb handler later.