linking works, db wrapper
Linking is done by the m3 build system (instead of rust directly). Currently we only use it to call
a simple test function from the dbwrapper library.

Also the library now contains what might be a minimal interface needed to call the db from Rust.
It compiles but is currently not connected to the code, so we'll see.