opaque wrapper for *db, build fails
At this point I need to pull up to the newer m3 version on the main branch. The reason is,
that for using c++ inside Rust both need to link against the same c/c++ base library (musl).
To make the Rust build do so, wee need an aditional argument ('std=True') which will (hopefully)
work in the newer version.