db connection works
It seems the connection from Rust to the level DB instance now works and we reimplemented the answer mechanism as in the app/bench/ycsb test case. In particular this means, we do not return the actual answer,
but only a vector of 0's having the length of answer.

Further todos are reacting to errors and closing the connection roperly, as well as adding timing.

In te benchmark case, timing was done before and after BD requests. However we're not interested in how the DB takes to answer, but in how long the whole loop from client to server and ack takes. So we need to
add timing to smoltcp_client/client.rs