basic statics
We want to measure the ip stack roundtrip time. So we can not use the
statics from the original ycsb-bench testcase but need to measure in the client. This has the slight disadvantage, that we don't know the
exact operation being executed. However we rather care for the size of requests and anwers anyways, which we can perfectly bookkeep in the client.