almost state local version, running
Now this implements basically the rewritten, state local version of the application. We'll still need to figure out how to handle waiting and semaphores properly (i.e. what to show or hide to/from Ohua).

Also we need to figure out how to handle file system mounting and handling of Activity cpabilities when we move forward to the isolated/
Activity spawning version.